[canvas] Restore Workpad Title/button to Home; fix mounting behavior (#103601)
