+++ x-pack/plugins/canvas/public/components/home/hooks/index.ts
@@ -9,7 +9,7 @@ export { useCloneWorkpad } from './use_clone_workpad';
 export { useCreateWorkpad } from './use_create_workpad';
 export { useDeleteWorkpads } from './use_delete_workpad';
 export { useDownloadWorkpad } from './use_download_workpad';
-export { useFindTemplates, useFindTemplatesOnMount } from './use_find_templates';
-export { useFindWorkpads, useFindWorkpadsOnMount } from './use_find_workpad';
+export { useFindTemplates } from './use_find_templates';
+export { useFindWorkpads } from './use_find_workpad';
 export { useImportWorkpad } from './use_upload_workpad';
 export { useCreateFromTemplate } from './use_create_from_template';
