+++ x-pack/plugins/canvas/public/components/home/loading.tsx
