+++ x-pack/plugins/canvas/public/components/toolbar/toolbar.component.tsx
@@ -93,43 +93,50 @@ export const Toolbar: FC<Props> = ({
     <div className="canvasToolbar hide-for-sharing">
       {activeTray !== null && <Tray done={() => setActiveTray(null)}>{trays[activeTray]}</Tray>}
       <div className="canvasToolbar__container">
-        <EuiFlexGroup alignItems="center" gutterSize="none" className="canvasToolbar__controls">
-          <EuiFlexItem grow={false}>
-            <RoutingButtonIcon
-              color="text"
-              to={getUrl(selectedPageNumber - 1)}
-              iconType="arrowLeft"
-              isDisabled={selectedPageNumber <= 1}
-              aria-label={strings.getPreviousPageAriaLabel()}
-            />
+        <EuiFlexGroup gutterSize="none" alignItems="center">
+          <EuiFlexItem grow={false} className="canvasToolbar__home">
+            {workpadName}
           </EuiFlexItem>
           <EuiFlexItem grow={false}>
-            <EuiButtonEmpty color="text" onClick={() => toggleTray('pageManager')}>
-              {strings.getPageButtonLabel(selectedPageNumber, totalPages)}
-            </EuiButtonEmpty>
+            <EuiFlexGroup alignItems="center" gutterSize="none" className="canvasToolbar__controls">
+              <EuiFlexItem grow={false}>
+                <RoutingButtonIcon
+                  color="text"
+                  to={getUrl(selectedPageNumber - 1)}
+                  iconType="arrowLeft"
+                  isDisabled={selectedPageNumber <= 1}
+                  aria-label={strings.getPreviousPageAriaLabel()}
+                />
+              </EuiFlexItem>
+              <EuiFlexItem grow={false}>
+                <EuiButtonEmpty color="text" onClick={() => toggleTray('pageManager')}>
+                  {strings.getPageButtonLabel(selectedPageNumber, totalPages)}
+                </EuiButtonEmpty>
+              </EuiFlexItem>
+              <EuiFlexItem grow={false}>
+                <RoutingButtonIcon
+                  color="text"
+                  to={getUrl(selectedPageNumber + 1)}
+                  iconType="arrowRight"
+                  isDisabled={selectedPageNumber >= totalPages}
+                  aria-label={strings.getNextPageAriaLabel()}
+                />
+              </EuiFlexItem>
+              <EuiFlexItem />
+              {elementIsSelected && isWriteable && (
+                <EuiFlexItem grow={false}>
+                  <EuiButtonEmpty
+                    color="text"
+                    iconType="editorCodeBlock"
+                    onClick={() => toggleTray('expression')}
+                    data-test-subj="canvasExpressionEditorButton"
+                  >
+                    {strings.getEditorButtonLabel()}
+                  </EuiButtonEmpty>
+                </EuiFlexItem>
+              )}
+            </EuiFlexGroup>
           </EuiFlexItem>
-          <EuiFlexItem grow={false}>
-            <RoutingButtonIcon
-              color="text"
-              to={getUrl(selectedPageNumber + 1)}
-              iconType="arrowRight"
-              isDisabled={selectedPageNumber >= totalPages}
-              aria-label={strings.getNextPageAriaLabel()}
-            />
-          </EuiFlexItem>
-          <EuiFlexItem />
-          {elementIsSelected && isWriteable && (
-            <EuiFlexItem grow={false}>
-              <EuiButtonEmpty
-                color="text"
-                iconType="editorCodeBlock"
-                onClick={() => toggleTray('expression')}
-                data-test-subj="canvasExpressionEditorButton"
-              >
-                {strings.getEditorButtonLabel()}
-              </EuiButtonEmpty>
-            </EuiFlexItem>
-          )}
         </EuiFlexGroup>
       </div>
     </div>
