+++ x-pack/plugins/canvas/public/components/toolbar/toolbar.scss
@@ -20,6 +20,11 @@
   }
 }
 
+.canvasToolbar__home {
+  padding: $euiSizeM 0 $euiSizeM $euiSizeL;
+  height: 100%;
+}
+
 .canvasToolbar__controls {
   padding: $euiSizeM;
   height: 100%;
